Rename package names in redis cache module
